Precompute row and column sums during packing and increase MR to 8
Precompute the row and column sums for each MR x NR tile while packing the LHS
and RHS respectively. This frees up registers for the matmul main loop, allowing
MR to be increased to 8 to match NR. As well as making use of the full width of
AVX registers, this also avoids the need for masking when loading.